using same Uint type for challenge in server and client

+++ apps/simple_follow/index.js
@@ -40,14 +40,14 @@ server.get("/", (req, res) => {
 
 server.get("/auth/account_creation_challenge", (req, res) => {
   handle = req.query.handle
-  challenge = Crypto.getRandomValues(new Uint32Array(1))[0];
+  challenge = Crypto.getRandomValues(new Uint8Array(1))[0];
   userId = sha256(handle);
   rpName = "Simple Follow";
   res.send({ challenge: challenge, user: { id: userId, name: handle, displayName: handle}, rpName: rpName })
 });
 
 server.get("/auth/login_challenge", (req, res) => {
-  challenge = Crypto.getRandomValues(new Uint32Array(1))[0];
+  challenge = Crypto.getRandomValues(new Uint8Array(1))[0];
   res.send({challenge: challenge})
 });
